pyproject.toml
[project]
name = "sinhala-ai-lyrics-video-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "moviepy==1.0.3",
    "openai==1.93.0",
    "opencv-python==4.11.0.86",
    "pillow==11.0.0",
    "python-dotenv==1.0.1",
    "python-telegram-bot==22.0",
]
